pyproject.toml
[tool.poetry]
name = "youtube-clipper"
version = "0.0.1"
description = ""
authors = ["Nick Proskurin <n1ckpr0s@yandex.ru>"]
packages = [{include = "youtube_clipper"}]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
attrs = ">=22.2.0"
colorama = ">=0.4.6"
python = ">=3.10"
srt = ">=3.5.3"
ttconv = ">=1.0.7"
whoosh = ">=2.7.4"
yt-dlp = ">=2023.10.13"

[tool.poetry.group.tests.dependencies]
pytest = ">=7.2.2"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = [
  "poetry-core",
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra"
testpaths = []

[tool.mypy]
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
strict_optional = true
